Fase 2/Evidencias Proyecto/Evidencias de sistema/Aplicación/pgf/apps/emergencies/urls.py
# apps/emergencies/urls.py
from rest_framework.routers import DefaultRouter
from .views import EmergenciaRutaViewSet

router = DefaultRouter()
router.register(r'', EmergenciaRutaViewSet, basename='emergencia')

urlpatterns = router.urls

